pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.79.0",
    "pinecone>=6.0.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.2",
]
